feat: Restructure locations of some type definitions to make them easier to import without large dependency trees
