+++ src/devices/casparCG.ts
@@ -19,6 +19,9 @@ import {
 	Mapping,
 	TimelineResolvedObjectExtended
 } from './mapping'
+import {
+	TimelineContentTypeCasparCg
+} from '../enums'
 
 import {
 	TimelineState,
@@ -51,17 +54,6 @@ export interface CasparCGOptions {
 	useScheduling?: boolean, // whether to use the CasparCG-SCHEDULE command to run future commands, or the internal (backwards-compatible) command queue
 	launcherHost: string,
 	launcherPort: string
-}
-export enum TimelineContentTypeCasparCg { //  CasparCG-state
-	VIDEO = 'video', // to be deprecated & replaced by MEDIA
-	AUDIO = 'audio', // to be deprecated & replaced by MEDIA
-	MEDIA = 'media',
-	IP = 'ip',
-	INPUT = 'input',
-	TEMPLATE = 'template',
-	HTMLPAGE = 'htmlpage',
-	ROUTE = 'route',
-	RECORD = 'record'
 }
 export class CasparCGDevice extends DeviceWithState<TimelineState> {
